requirements.txt
streamlit==1.46.0
python-dotenv==1.1.0
PyYAML==6.0.2
bcrypt==5.0.0
fpdf==1.7.2
numpy==1.26.2
certifi==2025.6.15
requests==2.32.3
urllib3==2.1.0
py3Dmol==2.5.1
stmol==0.0.9
chembl-webresource-client==0.10.9
pandas==2.1.3
rdkit==2025.3.5
joblib==1.3.2
PyPDF2==3.0.1
reportlab==4.4.4
umap-learn==0.5.5
plotly==6.4.0
scikit-learn==1.4.1.post1
matplotlib==3.8.3
networkx
pyvis
torch>=2.1.0
# Optional: deepchem for molecular modeling (heavy install via pip/conda)
# deepchem>=2.7.1
